% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uafrstat.R
\name{welcome}
\alias{welcome}
\title{Title for the welcome function}
\usage{
welcome()
}
\value{
a string.
}
\description{
Brief summary of the function- this function prints a welcoming statement.
}
\details{
If this function was more complex, you could include more details in this details section.
}
\examples{
welcome()
}
